requirements.txt
openai
pydantic
python-dotenv
loguru
spacy
tenacity
tqdm
groq
